requirements.txt
streamlit
camel-ai
camel-ai[rag]
unstructured[md]